Create header files for engine types and globals
All other files have been updated to use these headers as appropriate
