+++ src/tick.cpp
@@ -4,10 +4,11 @@
 
 #include "tick.hpp"
 
-#include "units.hpp"
 #include "threads.hpp"
 
 #include <unordered_map>
+
+#include "engine.hpp"
 
 TickHandler::TickHandler() : TickHandler(new ThreadPool(5)) {
 	this->createdPool = true;
